requirements.txt
boto3==1.34.34
requests==2.31.0
pandas==2.1.4
streamlit==1.32.0
snowflake-connector-python[pandas]==3.7.0
plotly==5.18.0
python-dotenv